+++ lib/src/lib/nav-list/components/icon/icon.component.ts
@@ -1,11 +1,12 @@
-import { ChangeDetectionStrategy, Component, input, ViewEncapsulation } from '@angular/core';
+import { ChangeDetectionStrategy, Component, input } from '@angular/core';
 import { Icon } from '../../models/nav-item.model';
 import { MatIcon } from '@angular/material/icon';
 
+/**
+ * Shows an Icon config as a Material icon or a registered SVG icon
+ */
 @Component({
   selector: 'lab900-icon',
-  styleUrls: ['./icon.component.scss'],
-  encapsulation: ViewEncapsulation.None,
   changeDetection: ChangeDetectionStrategy.OnPush,
   template: ` @if (icon().name) {
       <mat-icon>{{ icon().name }}</mat-icon>